ml/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
flatbuffers==23.5.26
idna==3.4
libclang==16.0.6
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
protobuf==4.25.0
pyasn1==0.5.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tensorboard-data-server==0.7.2
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wrapt==1.14.1
zipp==3.17.0
tensorflow